Refactored and added producer's classes
