+++ src/main/java/it/pagopa/interop/probing/eservice/operations/rest/ProducerController.java
@@ -6,17 +6,17 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.RestController;
 import it.pagopa.interop.probing.eservice.operations.api.ProducersApi;
 import it.pagopa.interop.probing.eservice.operations.dtos.SearchProducerNameResponse;
-import it.pagopa.interop.probing.eservice.operations.service.EserviceService;
+import it.pagopa.interop.probing.eservice.operations.service.ProducerService;
 
 @RestController
 public class ProducerController implements ProducersApi {
 
   @Autowired
-  EserviceService eserviceService;
+  ProducerService producerService;
 
   @Override
   public ResponseEntity<List<SearchProducerNameResponse>> getEservicesProducers(
       String producerName) {
-    return ResponseEntity.ok(eserviceService.getEservicesProducers(producerName));
+    return ResponseEntity.ok(producerService.getEservicesProducers(producerName));
   }
 }
