+++ src/main/java/it/pagopa/interop/probing/eservice/operations/util/constant/ProjectConstants.java
@@ -2,11 +2,13 @@ package it.pagopa.interop.probing.eservice.operations.util.constant;
 
 public class ProjectConstants {
 
-	public static final String ESERVICE_NAME_FIELD = "eserviceName";
-	public static final String ESERVICE_NAME_NATIVE_FIELD = "eservice_name";
+  public static final String ESERVICE_NAME_FIELD = "eserviceName";
+  public static final String ESERVICE_NAME_NATIVE_FIELD = "eservice_name";
 
-	private ProjectConstants() {
-		super();
-	}
-	
+  public static final String PRODUCER_NAME_FIELD = "producerName";
+
+  private ProjectConstants() {
+    super();
+  }
+
 }
