+++ src/test/java/it/pagopa/interop/probing/eservice/operations/integration/repository/EserviceRepositoryTest.java
@@ -1,9 +1,11 @@
 package it.pagopa.interop.probing.eservice.operations.integration.repository;
 
+import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import java.time.OffsetTime;
 import java.time.ZoneOffset;
+import java.util.List;
 import java.util.NoSuchElementException;
 import java.util.Optional;
 import java.util.UUID;
@@ -13,8 +15,10 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.boot.test.autoconfigure.orm.jpa.TestEntityManager;
+import org.springframework.data.domain.PageRequest;
 import it.pagopa.interop.probing.eservice.operations.dtos.EserviceStateBE;
 import it.pagopa.interop.probing.eservice.operations.dtos.EserviceTechnology;
+import it.pagopa.interop.probing.eservice.operations.dtos.SearchProducerNameResponse;
 import it.pagopa.interop.probing.eservice.operations.model.Eservice;
 import it.pagopa.interop.probing.eservice.operations.repository.EserviceRepository;
 
@@ -61,5 +65,31 @@ class EserviceRepositoryTest {
       queryResult.get();
     }, "There should be no e-service with eserviceId " + wrongVersionId + "and versionId "
         + wrongVersionId);
+  }
+
+  @Test
+  @DisplayName("when a valid producer name is provided, then the method should return a non-empty list")
+  void testGetEservicesProducers_whenGivenValidProducerName_thenReturnsEmptyList() {
+    List<SearchProducerNameResponse> resultGetEservicesProducers =
+        eserviceRepository.getEservicesProducers("producer1".toUpperCase(), PageRequest.of(0, 10));
+    assertEquals(1, resultGetEservicesProducers.size(),
+        "the method should return a non-empty list");
+  }
+
+  @Test
+  @DisplayName("when a producer name not saved on db is provided, then the method should return an empty list")
+  void testGetEservicesProducers_whenGivenProducerNameNotStored_thenReturnsEmptyList() {
+    List<SearchProducerNameResponse> resultGetEservicesProducers = eserviceRepository
+        .getEservicesProducers("Producer-To-Not-Found".toUpperCase(), PageRequest.of(0, 10));
+    assertEquals(0, resultGetEservicesProducers.size(), "the method should return an empty list");
+  }
+
+  @Test
+  @DisplayName("when a substring of a valid producer name is provided, then the method should return a non-empty list")
+  void testGetEservicesProducers_whenGivenPartialProducerName_thenReturnsNonEmptyList() {
+    List<SearchProducerNameResponse> resultGetEservicesProducers =
+        eserviceRepository.getEservicesProducers("pro".toUpperCase(), PageRequest.of(0, 10));
+    assertEquals(1, resultGetEservicesProducers.size(),
+        "the method should return a non-empty list");
   }
 }
