+++ src/test/java/it/pagopa/interop/probing/eservice/operations/integration/repository/EserviceViewRepositoryTest.java
@@ -3,7 +3,6 @@ package it.pagopa.interop.probing.eservice.operations.integration.repository;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import java.time.OffsetDateTime;
-import java.util.List;
 import java.util.UUID;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -12,11 +11,9 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.boot.test.autoconfigure.orm.jpa.TestEntityManager;
 import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort;
 import org.springframework.data.jpa.domain.Specification;
 import it.pagopa.interop.probing.eservice.operations.dtos.EserviceStateBE;
-import it.pagopa.interop.probing.eservice.operations.dtos.SearchProducerNameResponse;
 import it.pagopa.interop.probing.eservice.operations.model.view.EserviceView;
 import it.pagopa.interop.probing.eservice.operations.repository.EserviceViewRepository;
 import it.pagopa.interop.probing.eservice.operations.repository.specs.EserviceViewSpecs;
@@ -67,32 +64,6 @@ class EserviceViewRepositoryTest {
 
     assertNotNull(resultFindAll);
     assertEquals(0, resultFindAll.getTotalElements());
-  }
-
-  @Test
-  @DisplayName("when a valid producer name is provided, then the method should return a non-empty list")
-  void testGetEservicesProducers_whenGivenValidProducerName_thenReturnsEmptyList() {
-    List<SearchProducerNameResponse> resultGetEservicesProducers =
-        repository.getEservicesProducers("Producer Name".toUpperCase(), PageRequest.of(0, 10));
-    assertEquals(1, resultGetEservicesProducers.size(),
-        "the method should return a non-empty list");
-  }
-
-  @Test
-  @DisplayName("when a producer name not saved on db is provided, then the method should return an empty list")
-  void testGetEservicesProducers_whenGivenProducerNameNotStored_thenReturnsEmptyList() {
-    List<SearchProducerNameResponse> resultGetEservicesProducers = repository
-        .getEservicesProducers("Producer-To-Not-Found".toUpperCase(), PageRequest.of(0, 10));
-    assertEquals(0, resultGetEservicesProducers.size(), "the method should return an empty list");
-  }
-
-  @Test
-  @DisplayName("when a substring of a valid producer name is provided, then the method should return a non-empty list")
-  void testGetEservicesProducers_whenGivenPartialProducerName_thenReturnsNonEmptyList() {
-    List<SearchProducerNameResponse> resultGetEservicesProducers =
-        repository.getEservicesProducers("pro".toUpperCase(), PageRequest.of(0, 10));
-    assertEquals(1, resultGetEservicesProducers.size(),
-        "the method should return a non-empty list");
   }
 
 }
